Rename ReportType to CoverageReportType and move it to test.py
